requirements.txt
Django==3.2.15
djangorestframework==3.13.1
requests==2.26.0
python-dotenv==0.19.0
drf-yasg==1.20.0
djangorestframework-simplejwt==4.8.0